Make the published spec work with Apigee
Apigee SmartDocs is not as standards-compliant as Swagger/reference tooling.

It cannot handle external references, so we had to resolve those: `speccy resolve` command.

The resolved files still had *internal* fragment references which contained URLencoded characters - this had to be fixed: `-i` flag on speccy resolve command

Finally the output was converted to JSON using `js-yaml`.

Do these conversions yourself and compare to get a feel for the changes.