Change fuzzy match search so that it tests supported behaviour
Searching on historic date of births was never a supported feature in PDS FHIR nor in PDS HL7. This test worked because of a bug in PDS *not* filtering out stale indexes when searching on the inverted index. With the fix implemented in SpineII under SPINEDEM-3846, this test will fail, as it should have done from the beginning. It is now changed to a field on which one can search historic values, family name.